% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountryPopulation.R
\name{CountryPopulation}
\alias{CountryPopulation}
\title{A function utilizing the WorldPopulation data.frame to create a graphical representation depicting the population trend of any chosen country across time.}
\usage{
CountryPopulation(country, title = NULL)
}
\arguments{
\item{country}{Country name}

\item{title}{optional. Provide a title for the graph}
}
\value{
return the graph
}
\description{
A function utilizing the WorldPopulation data.frame to create a graphical representation depicting the population trend of any chosen country across time.
}
\examples{
CountryPopulation('Ghana')
CountryPopulation('Ghana', 'Time series data representing the population of Ghana.') Use the title instead of the Country name
}
