% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocWorldCup.R
\docType{data}
\name{World_Cup}
\alias{World_Cup}
\title{FIFA World Cup Statistics (1930 - 2022)}
\format{
a tibble containing data for 22 FIFA World Cup tournaments
\describe{
  \item{Year}{The year of the FIFA World Cup.}
  \item{Host}{The country hosting the FIFA World Cup tournament.}
  \item{Matches}{Total number of matches played in the tournament.}
  \item{Totalattendance}{Total attendance across all matches in the tournament.}
  \item{Averageattendance}{Average attendance per match in the tournament.}
}
}
\source{
\url{https://en.wikipedia.org/wiki/FIFA_World_Cup}
}
\usage{
World_Cup
}
\description{
FIFA World Cup Statistics (1930 - 2022)
}
\keyword{datasets}
