% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DocWorldPopulation.R
\docType{data}
\name{WorldPopulation}
\alias{WorldPopulation}
\title{The UN World Population dataset}
\format{
a tibble containing population statistics for 235 countries across 72 years
\describe{
  \item{CountryName}{Name of the country.}
  \item{`1950` to `2020`}{Population count for each respective year.}
}
}
\source{
\url{https://www.unfpa.org/data/world-population-dashboard}
}
\usage{
WorldPopulation
}
\description{
The UN World Population dataset
}
\keyword{datasets}
